typelanguage: split Rust backend into a separate module
